Create README.md
The Global Pi dollar app will enable the purchase of PiUSD with Pi, and the amount of Pi resulting from this purchase will be use to purchase stocks, commodities or gold within the Pi ecosystem to serve as reserve to tether and peg the PiUSD in 1:1 to the value of 1 USD. The PiUSD can then be use even on cryptocurrency exchanges and trading platforms to continue spreading and increasing awareness of Pi and Pi Network’s utility without compromising the value of Pi coin through trading for speculative price gain. The PiUSD will become the biggest competitor for to compete with other stable coins such as USDT and USDC, due to the wide spread of the Pi Network’s globally.